War y A+ extra

+++ Item 6 (A+ extra)/Acme CnG/src/main/webapp/views/actor/view.jsp
@@ -37,7 +37,7 @@
 <jstl:out value="${actor.phone}" />
 <br>
 <jstl:if test="${!isAdmin and myProfile}">
-	<a href="actor/delete.do"><spring:message code="actor.delete"/></a>
+	<a href="actor/delete.do" onclick="return confirm('<spring:message code="confirm.delete" />')" ><spring:message code="actor.delete"/></a>
 </jstl:if>
 <h2><spring:message  code="actor.comments" />:</h2>
 <jstl:if test="${!unBannedComments.isEmpty()}">
